🚚 Move and rename properties component
