+++ apps/cookbook/src/app/shared/api-description/api-description-properties/api-description-properties.component.html
