+++ apps/cookbook/src/app/showcase/avatar-showcase/avatar-showcase.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
+import { ShowcaseProperty } from '~/app/shared/api-description/api-description-properties/api-description-properties.component';
 
 import { Sizes } from '@kirbydesign/designsystem';
