+++ apps/cookbook/src/app/showcase/button-showcase/button-showcase.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
+import { ShowcaseProperty } from '~/app/shared/api-description/api-description-properties/api-description-properties.component';
 declare var require: any;
 
 @Component({
