+++ apps/cookbook/src/app/showcase/form-field-showcase/form-field-showcase.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ShowcaseEvent } from '~/app/shared/api-description/api-description-events/api-description-events.component';
 import { ShowcaseMethod } from '~/app/shared/api-description/api-description-methods/api-description-methods.component';
-import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
+import { ShowcaseProperty } from '~/app/shared/api-description/api-description-properties/api-description-properties.component';
 
 import { InputSize } from '@kirbydesign/designsystem';
