+++ apps/cookbook/src/app/showcase/showcase.module.ts
@@ -8,9 +8,9 @@ import { KirbyModule } from '@kirbydesign/designsystem';
 import { IphoneModule } from '../iphone/iphone.module';
 import { ShowcaseEventsComponent } from '../shared/api-description/api-description-events/api-description-events.component';
 import { ShowcaseMethodsComponent } from '../shared/api-description/api-description-methods/api-description-methods.component';
+import { ShowcasePropertiesComponent } from '../shared/api-description/api-description-properties/api-description-properties.component';
 import { CodeViewerComponent } from '../shared/code-viewer/code-viewer.component';
 import { ExampleViewerComponent } from '../shared/example-viewer/example-viewer.component';
-import { ShowcasePropertiesComponent } from '../shared/showcase-properties/showcase-properties.component';
 
 import { DividerShowcaseComponent } from './divider-showcase/divider-showcase.component';
 import { COMPONENT_DECLARATIONS, COMPONENT_EXPORTS, COMPONENT_IMPORTS } from './showcase.common';
